+++ ReleaseNotes.md
@@ -114,3 +114,4 @@ Please change your configuration according to [App.xaml](https://github.com/spie
   * Improved loading of preview images
   * Smoother scrolling in directories and files list
 * New `SaveFileControl.ForceFileExtensionOfFileFilter` property to enforce a file extension of the selected filter
+* Better support of dark theme in file system dialogs